Move cached observables
